assignment1-vector-search/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.0
pgvector==0.2.5
numpy==1.26.3
jupyter==1.0.0
ipykernel==6.29.0
